requirements.txt
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.2.0
